Font size combobox does not show fractional font sizes anymore, displays font information resolved with QFontInfo

+++ src/textedit.cpp
@@ -803,7 +803,7 @@ void TextEdit::SetFontFamily(QString family) {
 	applyCharFormatting(f);
 }
 
-void TextEdit::SetFontSize(qreal size) {
+void TextEdit::SetFontSize(int size) {
 	QTextCharFormat f;
 	f.setFontPointSize(size);
 	applyCharFormatting(f);
